requirements.txt
google-adk
python-gitlab
python-dotenv
requests